% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_metrics_calculation_helper.R
\name{boot_quantile}
\alias{boot_quantile}
\title{Bootstrap a quantile}
\usage{
boot_quantile(v, q, B = 1000)
}
\arguments{
\item{v}{A numeric vector.}

\item{q}{Numeric. Quantile to estimate, between 0 and 1.}

\item{B}{Integer. Number of bootstrap replicates.}
}
\value{
A numeric vector of bootstrap quantile estimates.
}
\description{
Estimates a quantile by bootstrap resampling a numeric vector.
}
