% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_metrics_calculation_helper.R
\name{safe_cor}
\alias{safe_cor}
\title{Calculate a correlation with safety checks}
\usage{
safe_cor(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A numeric value.
}
\description{
Returns the Pearson correlation between two numeric vectors. If either vector
has fewer than two values or zero variance, the function returns \code{NA_real_}.
}
