% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_metrics_calculation_helper.R
\name{safe_mean}
\alias{safe_mean}
\title{Calculate a mean with empty-vector protection}
\usage{
safe_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value.
}
\description{
Returns the mean of a numeric vector while ignoring missing values. If the
input vector has length zero, the function returns \code{NA_real_}.
}
