% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{species_list}
\alias{species_list}
\title{List of All Fig Wasp Species Observed on \emph{Ficus racemosa}}
\format{
A character vector of length 6.
}
\usage{
data(species_list)
}
\description{
This character vector contains the standardized names of all fig wasp species
observed on \emph{Ficus racemosa}, used in the fig wasp community simulator.
Each species name is formatted using an underscore to join the genus and species,
e.g., `Genus_species`. This vector defines the species order used in all parameter
lists (e.g., `entry_mu`, `fecundity_mean`, etc.), and it serves as the default
species pool for simulations.
}
\examples{
data(species_list)
print(species_list)
# [1] "Sycophaga_testacea"   "Apocrypta_sp"
#     "Sycophaga_mayri"     "Ceratosolen_sp"
#     "Sycophaga_agraensis" "Apocrypta_westwoodi"

}
\seealso{
\code{\link{parameter_list_default}} for parameters corresponding to each species.
}
\keyword{datasets}
